requirements.txt
tree-sitter>=0.23.0
tree-sitter-python>=0.23.0
tree-sitter-javascript>=0.23.0
tree-sitter-java>=0.23.0
tree-sitter-go>=0.23.0
tree-sitter-ruby>=0.23.0
tree-sitter-php>=0.23.0
tree-sitter-c>=0.23.0
tree-sitter-rust>=0.23.0
tree-sitter-cpp>=0.23.0
tree-sitter-c-sharp>=0.23.0
tree-sitter-typescript>=0.23.0
PyYAML>=6.0
